requirements.txt
faster-whisper==1.0.3
google-cloud-secret-manager>=2.20.2
google-cloud-storage>=2.17.0
librosa==0.10.2.post1
numpy<1.24
pyyaml==6.0.2
resemblyzer==0.1.3
scikit-learn==1.5.1
soundfile==0.12.1
torch==2.3.1
torchaudio==2.3.1
tqdm==4.66.4
webrtcvad==2.0.10